pluginCenterLogin now requires permissions & extract banner to its own file
