+++ scm-ui/ui-webapp/public/locales/de/admin.json
@@ -89,7 +89,7 @@
     "myCloudogu": {
       "login": {
         "button": {
-          "label": "Verbinden mit mycloudogu"
+          "label": "Verbinden mit"
         },
         "description": "Verbinden Sie Ihren SCM-Manager mit myCloudogu um besondere Plugins zu installieren. myCloudogu ist die Heimat der SCM-Manager Community, getragen von Maintainern des SCM-Managers. Sie haben noch kein Konto? Erstellen Sie während der Verbindung der SCM-Manager-Instanz kostenfrei ein myCloudogu-Konto."
       }
